- Added color select button to color filter in Batch Action Dialog  - Fixed some warnings - nothing critical  - Updated layout of Loading screen  - Removed duration tool from toolbar and deactivated hotkey (functionality is now contained by Pointer tool  - Added level of detail to easings rendering of fade bars  - Added new release -> v1.27

+++ Source/Control_ColorPicker.cpp
@@ -113,7 +113,7 @@ namespace MIDILightDrawer
 
 		// Use the smaller dimension to ensure the wheel stays circular and fits
 		int wheelSize = Math::Min(availableWidth, availableHeight);
-		_Radius_Wheel = (wheelSize / 2.0f);
+		_Radius_Wheel = ((float)wheelSize / 2.0f);
 		float innerRadius = _Radius_Wheel - RING_WIDTH; // Maintain exact 10-pixel ring width
 
 		if (wheelSize <= 0) return;
@@ -257,8 +257,8 @@ namespace MIDILightDrawer
 		_Label_Blue->Location	= Point(leftMargin, startY + 2 * (TEXT_BOX_HEIGHT + SPACING) + (TEXT_BOX_HEIGHT - _Label_Blue->Height) / 2);
 		_TextBox_Blue->Location = Point(textBoxX, startY + 2 * (TEXT_BOX_HEIGHT + SPACING));
 
-		_Label_Hex->Location	= Point(leftMargin, startY + 3.5 * (TEXT_BOX_HEIGHT + SPACING) + (TEXT_BOX_HEIGHT - _Label_Hex->Height) / 2);
-		_TextBox_Hex->Location	= Point(textBoxX, startY + 3.5 * (TEXT_BOX_HEIGHT + SPACING));
+		_Label_Hex->Location = Point(leftMargin, (int)(startY + 3.5 * (TEXT_BOX_HEIGHT + SPACING) + (TEXT_BOX_HEIGHT - _Label_Hex->Height) / 2));
+		_TextBox_Hex->Location = Point(textBoxX, (int)(startY + 3.5 * (TEXT_BOX_HEIGHT + SPACING)));
 	}
 	
 	void Control_ColorPicker::Update_Slider_Positions()
